+++ src/hooks/actions/use-conv-forward-as-attachment.ts
@@ -13,7 +13,7 @@ import { MIMETYPE_EML } from 'helpers/attachments';
 import { useMsgForwardAsAttachmentFn } from 'hooks/actions/use-msg-forward-as-attachment';
 import { ActionFn, UIActionDescriptor } from 'types/actions';
 import { UnsavedAttachment } from 'types/attachments';
-import { createEditBoard } from 'views/app/detail-panel/edit/legacyEditor/edit-view-board';
+import { createEditBoard } from 'views/app/detail-panel/edit/edit-view-board';
 
 type ConvForwardAsAttachmentAction = {
 	firstMessageId: string;
