+++ src/hooks/actions/use-conv-reply-all.tsx
@@ -11,7 +11,7 @@ import { useTranslation } from 'react-i18next';
 import { ConversationActionsDescriptors, EditViewActions } from 'constants/index';
 import { useMsgReplyAllFn } from 'hooks/actions/use-msg-reply-all';
 import { ActionFn, UIActionDescriptor } from 'types/actions';
-import { createEditBoard } from 'views/app/detail-panel/edit/legacyEditor/edit-view-board';
+import { createEditBoard } from 'views/app/detail-panel/edit/edit-view-board';
 
 type ConvReplyAllFunctionsParameter = {
 	firstMessageId: string;
