+++ src/views/app/detail-panel/edit/edit-view-board.tsx
