+++ src/views/app/detail-panel/edit/legacyEditor/edit-view.tsx
